initial setup is good

+++ calculateClusters.cu
@@ -10,7 +10,7 @@
 //(devicePoints, deviceClusters, returnClusterNum, numDimensions, numClusters, numPoints);
 
 //findClusters<<<1, numClusters>>>(devicePoints, deviceClusterAssignments, deviceClusters, numDimensions, numClusters, numPoints);
-__global__ void calculateClusterCenter(float *devicePoints, int *deviceClusterAssignments, float *deviceClusters, int numDimensions, int numClusters, int numPoints){
+__global__ void calculateClusterCenter(double *devicePoints, int *deviceClusterAssignments, double *deviceClusters, int numDimensions, int numClusters, int numPoints){
     int idx = blockIdx.x * blockDim.x + threadIdx.x;
 
     //for(int i = 0; i < 5; i++){
@@ -79,7 +79,7 @@ int main() {
     int numClusters = 3;
     int numPoints = 5;
     //float arrayPoints[];
-    std::vector<std::vector<float>> points = {
+    std::vector<std::vector<double>> points = {
         {8.8, 9.9, 0.0, 1.1},
         {2.1, 4.3, 6.5, 8.7},
         {4.4, 5.5, 6.6, 7.7},
@@ -87,7 +87,7 @@ int main() {
         {7.8, 7.9, 7.0, 7.1}
     };
     std::vector<int> clusterAssignmentsIn = {1,1,2,2,2};
-    std::vector<std::vector<float>> clusters = {
+    std::vector<std::vector<double>> clusters = {
         {3.2, 3.4, 7.6, 2.8},
         {4.4, 5.5, 6.6, 7.7},
         {8.8, 9.9, 0.0, 1.1},
@@ -96,12 +96,12 @@ int main() {
 
     int pointsLengthNeeded = numDimensions * numPoints;
     int clustersLengthNeeded = numDimensions * numClusters;
-    float pointsArray[pointsLengthNeeded];
+    double pointsArray[pointsLengthNeeded];
     int clusterAssignmentsArray[numPoints];
     //float clustersArray[clustersLengthNeeded];
-    float *devicePoints;
+    double *devicePoints;
     int *deviceClusterAssignments;
-    float *deviceClusters;
+    double *deviceClusters;
     
     for(int i = 0; i < numPoints; i++){
         for(int j = 0; j < numDimensions; j++){
@@ -123,18 +123,18 @@ int main() {
 
 
     //this stuff allocates the memory needed for the array
-    cudaMalloc((void**)&devicePoints, pointsLengthNeeded * sizeof(float));
+    cudaMalloc((void**)&devicePoints, pointsLengthNeeded * sizeof(double));
     cudaMalloc((void**)&deviceClusterAssignments, numPoints * sizeof(int));
-    cudaMalloc((void**)&deviceClusters, clustersLengthNeeded * sizeof(float));
+    cudaMalloc((void**)&deviceClusters, clustersLengthNeeded * sizeof(double));
 
     //yo basically this just copies the data from the host array to the device array to the position that is already allocated and it will be used
-    cudaMemcpy(devicePoints, pointsArray, pointsLengthNeeded * sizeof(float), cudaMemcpyHostToDevice);
+    cudaMemcpy(devicePoints, pointsArray, pointsLengthNeeded * sizeof(double), cudaMemcpyHostToDevice);
     cudaMemcpy(deviceClusterAssignments, clusterAssignmentsArray, numPoints * sizeof(int), cudaMemcpyHostToDevice);
 
     //this is one block of 10 threads - try multiple blocks and just get an understanding of it more than high level
     calculateClusterCenter<<<1, numClusters>>>(devicePoints, deviceClusterAssignments, deviceClusters, numDimensions, numClusters, numPoints);
-    float clustersNewCenter[clustersLengthNeeded];
-    cudaMemcpy(clustersNewCenter, deviceClusters, clustersLengthNeeded * sizeof(float), cudaMemcpyDeviceToHost);
+    double clustersNewCenter[clustersLengthNeeded];
+    cudaMemcpy(clustersNewCenter, deviceClusters, clustersLengthNeeded * sizeof(double), cudaMemcpyDeviceToHost);
     
     for(int i = 0; i < numClusters; i++){
         std::cout << "Cluster #" << i << ": ";
